Get title from snippet where possible
It's not uncommon for posts to have no title. At present time,
Fluentflame reader does something incredibly stupid: write into
the database what the title should be based on your current
localisation. This incredibly stupid.

Let's not do that. We first guess what a valid title should be
from a snippet, and we use that. If no snippet is available, then
we go to the untitled text.

Also update the CSS so we can wrap text correctly.